Move back_insert_iterator into iterator.h
